v0.98.7:palm_tree:
