+++ config.py
@@ -113,3 +113,9 @@ portable_execute_file_size=20480 #默认20M大小
 #DNSLOG配置
 #########################################################################
 domain_name_system_address="dnslog.ascotbe.com" #用户用来接收数据的DNSLOG域名
+
+
+#########################################################################
+#Nist数据配置
+#########################################################################
+nist_update_job_time=7200#工作间隔，网站每2小时更新一次数据
