+++ docs/Documentation/UpDataLog.md
@@ -2559,10 +2559,19 @@ v0.98.5🌴
 ```
 
 
->  2021.4.1
+>  2021.4.11
 
 ```
 v0.98.6🌴
 添加DNSLOG功能
 添加DNSLOG后端说明文档
+```
+
+
+>  2021.4.12
+
+```
+v0.98.7🌴
+添加Nist数据更新功能
+优化Nist文件分类
 ```
